Small additions

+++ BudgetingApp/src/main/java/budgetingapp/dao/UserDao.java
@@ -32,9 +32,10 @@ public class UserDao {
 
     /**
      * Add new User object to the database
+     *
      * @param user object to add
-     * @return true if 
-     * @throws SQLException 
+     * @return true if
+     * @throws SQLException
      */
     public String add(User user) throws SQLException {
         ps = db.prepareStatement("INSERT INTO user (firstname, lastname, username, password, balance) VALUES (?, ?, ?, ?, ?)");
@@ -43,7 +44,7 @@ public class UserDao {
             ps.setString(2, user.getLastName());
             ps.setString(3, user.getUsername());
             ps.setString(4, user.getPassword());
-            ps.setInt(5, user.getBalance());
+            ps.setDouble(5, user.getBalance());
             ps.execute();
             return "Success";
         } catch (SQLException e) {
@@ -53,18 +54,22 @@ public class UserDao {
             return "Error";
         }
     }
-    
+
     public String delete(String username) throws SQLException {
-        ps = db.prepareStatement("DELETE FROM user WHERE username=(?)");
-        try {
-            ps.setString(1, username);
-            ps.execute();
-            return "Success";
-        } catch (SQLException e){
-            return "Failure";
+        if (get(username) != null) {
+            ps = db.prepareStatement("DELETE FROM user WHERE username=(?)");
+            try {
+                ps.setString(1, username);
+                ps.execute();
+                return "Success";
+            } catch (SQLException e) {
+                return "Failure";
+            }
+        } else {
+            return "Doesnt exist";
         }
     }
-    
+
     public User get(String username) throws SQLException {
         User user = null;
         ps = db.prepareStatement("SELECT * FROM user WHERE username=(?)");
@@ -73,18 +78,18 @@ public class UserDao {
             rs = ps.executeQuery();
             while (rs.next()) {
                 user = new User(
-                rs.getString("firstname"),
-                rs.getString("lastname"),
-                rs.getString("username"),
-                rs.getString("password"),
-                rs.getInt("balance"));
+                        rs.getString("firstname"),
+                        rs.getString("lastname"),
+                        rs.getString("username"),
+                        rs.getString("password"),
+                        rs.getInt("balance"));
             }
         } catch (SQLException e) {
-            
+
         }
         return user;
     }
-    
+
     public int getIdByUser(User user) throws SQLException {
         int id = 0;
         ps = db.prepareStatement("SELECT id FROM user WHERE username=(?)");
@@ -95,11 +100,31 @@ public class UserDao {
                 id = rs.getInt("id");
             }
         } catch (SQLException e) {
-            
+
         }
         return id;
     }
     
+    public User getById(int id) throws SQLException {
+        User user = null;
+        ps = db.prepareStatement("SELECT * FROM user WHERE id=(?)");
+        try {
+            ps.setInt(1, id);
+            rs = ps.executeQuery();
+            while (rs.next()) {
+                user = new User(
+                        rs.getString("firstname"),
+                        rs.getString("lastname"),
+                        rs.getString("username"),
+                        rs.getString("password"),
+                        rs.getInt("balance"));
+            }
+        } catch (SQLException e) {
+            
+        }
+        return user;
+    }
+
     public String login(String username, String password) throws SQLException {
         ps = db.prepareStatement("SELECT * FROM user WHERE username=(?) AND password=(?)");
         try {
@@ -113,6 +138,18 @@ public class UserDao {
             return "Failure";
         }
         return "Failure";
+    }
+    
+    public String updateBalance(int userId, double newAmount) throws SQLException {
+        ps = db.prepareStatement("UPDATE user SET balance=(?) WHERE id=(?)");
+        try {
+            ps.setDouble(1, newAmount);
+            ps.setInt(2, userId);
+            ps.executeUpdate();
+            return "Success";
+        } catch (SQLException e) {
+            return "Failure";
+        }
     }
 
 }
